refactor: move DB seed script from tests/e2e to scripts
The seed script doubles as a dev tool (make dev-db-seed), so scripts/
alongside admin.ts and run-migrations.ts is a more fitting home than
tests/e2e. E2E tests still consume it via the @/ alias.
